+++ dataforge-context/src/commonMain/kotlin/space/kscience/dataforge/context/Context.kt
@@ -9,9 +9,7 @@ import kotlinx.serialization.modules.SerializersModule
 import kotlinx.serialization.modules.overwriteWith
 import kotlinx.serialization.modules.plus
 import space.kscience.dataforge.meta.*
-import space.kscience.dataforge.misc.DFExperimental
 import space.kscience.dataforge.misc.Named
-import space.kscience.dataforge.misc.ThreadSafe
 import space.kscience.dataforge.names.Name
 import space.kscience.dataforge.provider.Provider
 import kotlin.coroutines.CoroutineContext
@@ -81,17 +79,53 @@ public open class Context internal constructor(
     private val childrenContexts = HashMap<Name, Context>()
 
     /**
-     * Build and register a new child context.
+     * Build and register a new child context. If context with the same name, properties and plugins already exists, reuse it.
+     * Throw an exception if context with the same name exists but is different.
      * @param name the relative (tail) name of the new context. If null, use context hash code as a marker.
      */
-    @OptIn(DFExperimental::class)
-    @ThreadSafe
-    public fun buildContext(name: Name? = null, block: ContextBuilder.() -> Unit = {}): Context {
-        val existing = name?.let { childrenContexts[name] }
-        if (existing != null) error("Context $name already exists in $this")
-        return ContextBuilder(this, name).apply(block).build().also {
+    public fun buildContext(
+        name: String? = null,
+        block: ContextBuilder.() -> Unit = {},
+    ): Context = ContextBuilder(this, name).apply(block).build().also {
+        val existing = childrenContexts[it.name]
+        if (existing != null) {
+            if (equals(existing, it)) {
+                //reuse existing context if it is the same
+                return existing
+            } else {
+                error("Context with name ${it.name} already exists but has different properties and plugins")
+            }
+        } else {
             childrenContexts[it.name] = it
         }
+    }
+
+    /**
+     * Check if the current context already has plugins and properties matching the given block.
+     * If it does, return it, otherwise create a new child context ensured to have them
+     */
+    public fun deriveContext(
+        block: ContextBuilder.() -> Unit = {},
+    ): Context {
+        val builder = ContextBuilder(this, meta = properties).apply(block)
+
+        val requiresFork = !Meta.equals(properties, builder.meta)
+                || builder.factories.any { (factory, meta) ->
+            val loaded = plugins[factory.tag]
+            loaded == null || loaded.meta != meta
+        }
+
+        if (!requiresFork) return this
+
+        // Search for existing child that adheres to restrictions
+        childrenContexts.values.find { child ->
+            Meta.equals(child.properties, builder.meta) && builder.factories.all { (factory, meta) ->
+                val loaded = child.plugins[factory.tag]
+                loaded != null && loaded.meta == meta
+            }
+        }?.let { return it }
+
+        return buildContext(block = block)
     }
 
     /**
@@ -111,7 +145,7 @@ public open class Context internal constructor(
     }
 
     override fun toString(): String {
-        val parentString = if(parent == Global) "" else ", parent=$parent"
+        val parentString = if (parent == Global) "" else ", parent=$parent"
         return "Context(name=$name$parentString)"
     }
 
@@ -150,7 +184,23 @@ public open class Context internal constructor(
 
     public companion object {
         public const val PROPERTY_TARGET: String = "context.property"
+
+        internal fun equals(c1: Context, c2: Context): Boolean =
+            c1.properties == c2.properties &&
+                    c1.plugins.tags == c2.plugins.tags &&
+                    c1.plugins.all { c2.plugins[it.tag]?.meta == it.meta }
+
     }
+}
+
+/**
+ * Fetch a plugin with given meta from the context. If the plugin (with given meta) is already registered, it is returned.
+ * Otherwise, new child context with the plugin is created. In the later case the context could be retrieved from the plugin.
+ */
+public inline fun <reified T : Plugin> Context.request(factory: PluginFactory<T>, meta: Meta? = null): T {
+    return deriveContext {
+        plugin(factory, meta ?: Meta.EMPTY)
+    }.plugins[factory]!!
 }
 
 /**