+++ dataforge-context/src/commonMain/kotlin/space/kscience/dataforge/context/Global.kt
@@ -4,7 +4,6 @@ import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.Job
 import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.names.Name
-import space.kscience.dataforge.names.parseAsName
 import kotlin.coroutines.CoroutineContext
 import kotlin.native.concurrent.ThreadLocal
 
@@ -20,5 +19,11 @@ private object GlobalContext : Context(Name.of("GLOBAL"), null, emptySet(), Meta
 
 public val Global: Context get() = GlobalContext
 
-public fun Context(name: String? = null, block: ContextBuilder.() -> Unit = {}): Context =
-    Global.buildContext(name?.parseAsName(), block)
+/**
+ * Create a new context with given configuration [block]. Could reuse an existing context if [name] is null and
+ * the created context in identical to the existing one.
+ */
+public fun Context(
+    name: String? = null,
+    block: ContextBuilder.() -> Unit = {}
+): Context = Global.buildContext(name, block = block)