+++ dataforge-workspace/src/commonMain/kotlin/space/kscience/dataforge/workspace/WorkspaceBuilder.kt
@@ -114,8 +114,8 @@ public class WorkspaceBuilder(
     /**
      * Define a context for the workspace
      */
-    public fun context(contextName: Name? = null, block: ContextBuilder.() -> Unit = {}) {
-        this.context = parentContext.buildContext(contextName, block)
+    public fun context(contextName: String? = null, block: ContextBuilder.() -> Unit = {}) {
+        this.context = parentContext.buildContext(contextName, block = block)
     }
 
     /**